// src/types/enums.ts
export enum ButtonIconType {
  shoppingCartOutlined = "ShoppingCartOutlined",
  addShoppingCart = "AddShoppingCart",
  shoppingCart = "ShoppingCart",
  menu = "Menu",
  phone = "Phone",
  search = "Search",
  instagram = "Instagram",
  personOutline = "PersonOutline",
  chevronLeft = "ChevronLeft",
  chevronRight = "ChevronRight",
  firstPage = "FirstPage",
  lastPage = "LastPage",
  lockOutlined = "LockOutlined",
}

export enum Colors {
  blue = '#03B5E4',
  red = '#F5043E',
  grey = '#B6BFCC',
  whiteGrey = '#C9CDD5',
  yellow = '#FEDB5C',
  pink = '#FD6594',
  green = '#6FCF97',
  white = '#fff',
  light = 'F5F7FA',
  transparent = 'transparent',
  purple = '#947AFC',
  blackblue = '#193888',
  text= '#132C4A',
  disabledText = '#626D7D',
  darkYellow = '#F6C000'
}